# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "슬기로운 계약생활 API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/contracts": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "List Contracts",
        "description": "대시보드용 본인 계약 건 목록 (최신 생성 순).",
        "operationId": "list_contracts_api_contracts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ContractResponse"
                  },
                  "type": "array",
                  "title": "Response List Contracts Api Contracts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Create Contract",
        "operationId": "create_contract_api_contracts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/contracts/{contract_id}": {
      "get": {
        "tags": [
          "contracts"
        ],
        "summary": "Get Contract",
        "operationId": "get_contract_api_contracts__contract_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contracts"
        ],
        "summary": "Delete Contract",
        "operationId": "delete_contract_api_contracts__contract_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contracts/{contract_id}/situation": {
      "put": {
        "tags": [
          "contracts"
        ],
        "summary": "Put Situation",
        "description": "계약 상황 입력 (사용자 흐름 3단계). 재입력 시 덮어씀.",
        "operationId": "put_situation_api_contracts__contract_id__situation_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SituationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contracts/{contract_id}/registry-link": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Link Registry",
        "description": "모의 등기 데이터 연결 (2026-07-16 팀 합의 API). 규칙 엔진 교차검증에 사용.",
        "operationId": "link_registry_api_contracts__contract_id__registry_link_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistryLinkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contracts/{contract_id}/analysis-runs": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Start Analysis",
        "description": "최신 확인 완료 스냅샷으로 분석을 시작한다. 재분석마다 새 행이 쌓인다(이력).",
        "operationId": "start_analysis_api_contracts__contract_id__analysis_runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "List Analysis Runs",
        "description": "재분석 이력 (최신순).",
        "operationId": "list_analysis_runs_api_contracts__contract_id__analysis_runs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisRunSummary"
                  },
                  "title": "Response List Analysis Runs Api Contracts  Contract Id  Analysis Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/contracts/{contract_id}/analysis-runs/{analysis_run_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis Run",
        "description": "분석 상태·결과 폴링. completed면 result에 통합 AnalysisRunResult JSON이 담긴다.",
        "operationId": "get_analysis_run_api_contracts__contract_id__analysis_runs__analysis_run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "contract_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contract Id"
            }
          },
          {
            "name": "analysis_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisRunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisRunDetail": {
        "properties": {
          "analysis_run_id": {
            "type": "string",
            "title": "Analysis Run Id"
          },
          "input_snapshot_id": {
            "type": "string",
            "title": "Input Snapshot Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "analysis_run_id",
          "input_snapshot_id",
          "status",
          "created_at"
        ],
        "title": "AnalysisRunDetail",
        "description": "분석 실행 1회 — result는 completed 후 통합 AnalysisRunResult JSON, 그 전엔 null."
      },
      "AnalysisRunSummary": {
        "properties": {
          "analysis_run_id": {
            "type": "string",
            "title": "Analysis Run Id"
          },
          "input_snapshot_id": {
            "type": "string",
            "title": "Input Snapshot Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "analysis_run_id",
          "input_snapshot_id",
          "status",
          "created_at"
        ],
        "title": "AnalysisRunSummary",
        "description": "재분석 이력 목록 항목."
      },
...
      "ContractCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ContractCreateRequest"
      },
      "ContractResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "contract_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractType"
              },
              {
                "type": "null"
              }
            ]
          },
          "contract_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "deposit_paid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Paid"
          },
          "signed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed"
          },
          "move_in_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move In Date"
          },
          "balance_payment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Payment Date"
          },
          "is_proxy_contract": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Proxy Contract"
          },
          "registry_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registry Case Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "contract_type",
          "contract_stage",
          "deposit_paid",
          "signed",
          "move_in_date",
          "balance_payment_date",
          "is_proxy_contract",
          "registry_case_id",
          "created_at"
        ],
        "title": "ContractResponse"
      },
      "ContractStage": {
        "type": "string",
        "enum": [
          "계약금 입금 전",
          "서명 전",
          "계약 직후"
        ],
        "title": "ContractStage"
      },
      "ContractType": {
        "type": "string",
        "enum": [
          "전세",
          "보증부 월세",
          "일반 월세"
        ],
        "title": "ContractType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegistryLinkRequest": {
        "properties": {
          "case_id": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Case Id"
          }
        },
        "type": "object",
        "required": [
          "case_id"
        ],
        "title": "RegistryLinkRequest",
        "description": "모의 등기 연결 — data/sample/registry-records의 합성 사례 식별자."
      },
...
      "SituationRequest": {
        "properties": {
          "contract_type": {
            "$ref": "#/components/schemas/ContractType"
          },
          "contract_stage": {
            "$ref": "#/components/schemas/ContractStage"
          },
          "deposit_paid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposit Paid"
          },
          "signed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed"
          },
          "move_in_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Move In Date"
          },
          "balance_payment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Balance Payment Date"
          },
          "is_proxy_contract": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Proxy Contract"
          }
        },
        "type": "object",
        "required": [
          "contract_type",
          "contract_stage"
        ],
        "title": "SituationRequest",
        "description": "계약 상황 입력 (사용자 흐름 3단계). 필드 구성은 통합 ContractContext 기준.\n\ndeposit_paid 이하 5개는 기존 클라이언트·mock 호환을 위해 선택 입력(None 허용).\n분석 실행 시 ContractContext가 필수값을 최종 검증한다."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}